refactor: 상품 redux 분리
