+++ src/pages/ProductList/index.tsx
@@ -1,15 +1,20 @@
 import { useEffect } from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
-import { loadProductsAPI, ProductState, selectProductState } from '@/redux/modules/products';
+import { RootState } from '@/redux/store';
+import { loadProductsAPI } from '@/redux/modules/product/productThunk';
 
 import { ProductListContainer } from './styles';
+
+import { ProductState } from '@/types';
 
 import { Loader } from '@/components/@shared';
 import Product from '@/components/Product';
 
 function ProductList() {
-  const { productList, loading, error }: ProductState = useSelector(selectProductState);
+  const { productList, loading, error }: ProductState = useSelector(
+    (state: RootState) => state.products
+  );
   const dispatch = useDispatch();
 
   useEffect(() => {
