+++ src/redux/store.ts
@@ -6,7 +6,7 @@ import storage from 'redux-persist/lib/storage';
 
 import cartReducer from './modules/cart';
 import customerReducer from './modules/customer';
-import productsReducer from './modules/products';
+import productsReducer from './modules/product/productReducer';
 import snackBarReducer from './modules/snackBar';
 
 export type AppDispatch = typeof store.dispatch;
